requirements.txt
numpy
scipy
h5py
opencv-python
matplotlib
pandas
parmap
colormath
cupy-cuda113
statsmodels
sklearn